+++ app/ante/cosmos/min_price_test.go
@@ -5,7 +5,7 @@ import (
 	sdk "github.com/cosmos/cosmos-sdk/types"
 	banktypes "github.com/cosmos/cosmos-sdk/x/bank/types"
 	cosmosante "github.com/evmos/evmos/v11/app/ante/cosmos"
-	evmtypes "github.com/evmos/evmos/v11/x/evm/types"
+	"github.com/evmos/evmos/v11/utils"
 )
 
 var execTypes = []struct {
@@ -18,7 +18,7 @@ var execTypes = []struct {
 }
 
 func (suite *AnteTestSuite) TestMinGasPriceDecorator() {
-	denom := evmtypes.DefaultEVMDenom
+	denom := utils.BaseDenom
 	testMsg := banktypes.MsgSend{
 		FromAddress: "evmos1x8fhpj9nmhqk8z9kpgjt95ck2xwyue0ptzkucp",
 		ToAddress:   "evmos1dx67l23hz9l0k9hcher8xz04uj7wf3yu26l2yn",
@@ -114,6 +114,26 @@ func (suite *AnteTestSuite) TestMinGasPriceDecorator() {
 			},
 			false,
 			"provided fee < minimum global fee",
+			true,
+		},
+		{
+			"valid cosmos tx without specified fee with MinGasPrices = 10, gasPrice = 10",
+			func() sdk.Tx {
+				params := suite.app.FeeMarketKeeper.GetParams(suite.ctx)
+				params.MinGasPrice = sdk.NewDec(10)
+				err := suite.app.FeeMarketKeeper.SetParams(suite.ctx, params)
+				suite.Require().NoError(err)
+
+				txBuilder := suite.clientCtx.TxConfig.NewTxBuilder()
+				txBuilder.SetGasLimit(TestGasLimit)
+				// fee is nil
+				txBuilder.SetFeeAmount(nil)
+				err = txBuilder.SetMsgs(&testMsg)
+				suite.Require().NoError(err)
+				return txBuilder.GetTx()
+			},
+			false,
+			"fee not provided",
 			true,
 		},
 	}