+++ app/app_test.go
@@ -21,7 +21,6 @@ import (
 
 	"github.com/evmos/evmos/v11/encoding"
 	"github.com/evmos/evmos/v11/utils"
-	claimstypes "github.com/evmos/evmos/v11/x/claims/types"
 )
 
 func TestEvmosExport(t *testing.T) {
@@ -39,7 +38,7 @@ func TestEvmosExport(t *testing.T) {
 	acc := authtypes.NewBaseAccount(senderPrivKey.PubKey().Address().Bytes(), senderPrivKey.PubKey(), 0, 0)
 	balance := banktypes.Balance{
 		Address: acc.GetAddress().String(),
-		Coins:   sdk.NewCoins(sdk.NewCoin(claimstypes.DefaultParams().ClaimsDenom, sdk.NewInt(100000000000000))),
+		Coins:   sdk.NewCoins(sdk.NewCoin(utils.BaseDenom, sdk.NewInt(100000000000000))),
 	}
 
 	db := dbm.NewMemDB()