+++ tests/e2e/upgrade/queryexec.go
@@ -1,0 +1,19 @@
+package upgrade
+
+import (
+	"fmt"
+)
+
+// CreateModuleQueryExec creates a Evmos module query
+func (m *Manager) CreateModuleQueryExec(moduleName, subCommand, chainID string) (string, error) {
+	cmd := []string{
+		"evmosd",
+		"q",
+		moduleName,
+		subCommand,
+		fmt.Sprintf("--chain-id=%s", chainID),
+		"--keyring-backend=test",
+		"--log_format=json",
+	}
+	return m.CreateExec(cmd, m.ContainerID())
+}