+++ x/evm/types/dynamic_fee_tx.go
@@ -260,10 +260,19 @@ func (tx DynamicFeeTx) Validate() error {
 		}
 	}
 
-	if tx.GetChainID() == nil {
+	chainID := tx.GetChainID()
+
+	if chainID == nil {
 		return errorsmod.Wrap(
 			errortypes.ErrInvalidChainID,
 			"chain ID must be present on AccessList txs",
+		)
+	}
+
+	if !(chainID.Cmp(big.NewInt(9001)) == 0 || chainID.Cmp(big.NewInt(9000)) == 0) {
+		return errorsmod.Wrapf(
+			errortypes.ErrInvalidChainID,
+			"chain ID must be 9000 or 9001 on Evmos, got %s", chainID,
 		)
 	}